Add support for custom trace function, ApplicationTrace()
This change makes it possible to build OpENer in a way where the
application can define a callback function, ApplicationTrace(),
that handles all logging instead of the built-in defaults.  This
makes it possible to use, e.g., the UNIX syslog() API, and/or to
adjust log level dynamically at runtime.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>